# src/main/java/com/gordeev/bankcards/dto/user/UserStatusUpdateRequest.java
package com.gordeev.bankcards.dto.user;

import jakarta.validation.constraints.NotNull;

import java.util.UUID;

public record UserStatusUpdateRequest(
        @NotNull(message = "ID пользователя обязателен")
        UUID userId,

        @NotNull(message = "Статус enabled обязателен")
        boolean enabled
) {
}